fix(rust): invalidate cache before enabling MMU
By design, there is currently some overlap in memory of BL2 and BL31. By
the time we are running BL31 we have finished with the BL2 boot loader
stage and so usually this does not cause any problems.

However when enabling translation in pagetable.rs:init() we call
enable_mmu_direct_el3() which also enables caching at BL31 via the
SCTLR_EL3.C bit. Where there are dirty cache lines from previous boot
loader stages in the overlapping regions, these cache lines can now be
pulled in when loading memory.

This issue is particularly evident where the overlap is large due to the
increased image sizes of debug builds. As a result parts of the
overlapping BL31 stack can be corrupted with stale cache data from BL2.

Invalidate cache for the entire BL31 image in the BL31 entrypoint to
avoid this.

Previously TF-A has chosen to invalidate only the RW sections as it
assumes that BL2 and BL31 share the same linear or ID map for BL31 RO,
and that this area will not be modified (except in the case of PIE). As
we cannot make assumptions about the operation of previous BL stages
(for example, a 3rd party bootloader loading Rust BL31) the full image
is invalidated to be safe.

Change-Id: Ie67cb7525b4494657d5b51a5a1a8ced69f4bea5a
Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Co-authored-by: Olivier Deprez <olivier.deprez@arm.com>