build(rust): prevent 'make fvp/qemu' command from triggering builds
Currently, when "make fvp" or "make qemu" is executed, this triggers
a build of Rust BL31.

 * Change target dependencies so that this build is not triggered
   and only soft links re-directing to the already built binaries
   are created.

This can be done as the soft links are not strictly a part of the
build process and are only used when running the emulation software.
The BL31 binary should have already been built from the top level
directory before executing 'make fvp/qemu'

The dependencies for "make run" from the top level directory should
be enough to build the software before "make fvp" is executed.

Change-Id: I6c6e8659ff14c156042c2931d5c02c3c377da099
Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>