build(rust): separate BL31 and BL33 builds
Set clear dependencies in the build process of the top level
directory(for building FIP, for example, as only making FIP
dependent on bl31 and calling "make all" on the Rust side is
inconsistent with that's really happening and can lead to issues
when building).

Currently, the `make -C rust PLAT=fvp build` command will create a
`bl31.elf` file but will not create the Rust BL31 binary bl31.bin,
which is not what we expect.

 * Make it so that `make -C rust PLAT=fvp build` creates the
   bl31.bin

This is just temporary in the prototype, RF-A would not normally
be building any kind of BL33.

 * Separate the Rust logic in a rust/bl31.mk file so that it's
   easier to remove after we finish prototyping.

Change-Id: I6484e81464a343ba4d58cad9992dfd20508d1dfc
Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>