refactor(rust): implement accurate SMC dispatch
Previously we were passing all SMCs with Owning Entity Number (OEN) 4 to
the PSCI service, however this is not accurate. OEN 4 refers to Standard
Secure Services and covers a range of SMCs in addition to PSCI,
including FF-A and CCA.

Implement an owns() function for each service that specifies which range
of SMCs it handles. Using the helper macro owns! in services.rs can be
used to do this for standard use cases.

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>
Change-Id: Ib2e414fb7e133ed7d98a3d531e9ca43d0234d2c2