feat(rust): setup Realm world for FEAT_RME
Add initial support required for Realm Management Extension aka
FEAT_RME, part of Arm's Confidential Compute Architecture (CCA).

This commit does not yet boot an RMM payload in R-EL2, but adds the
infrastructure and stubbed out functions/SMC handlers to do so. See
https://trustedfirmware-a.readthedocs.io/en/latest/components/rmm-el3-comms-spec.html
for details of this interface.

The boot flow and context switching between Root -> Realm worlds is
handled in a later patch.

Build & run from the top level TF-A Makefile with:
> make PLAT=fvp RUST_FEATURES=rme DEBUG=1 run

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>
Signed-off-by: Andrew Walbran <qwandor@google.com>
Change-Id: Ic4f3b6fe75f0d28d7a16c2cddbfddf2a9011c1f0