feat(rust): start on secure world test framework
This adds two binaries: one to run in S-EL2 as a "fake SPMC" BL32, and
one to run in EL2 as BL33. The BL33 component will co-ordinate running
tests by sending FF-A messages to the BL32 component, but for now this
initial commit adds both binaries with some basic UART logs to show that
they have started.

Change-Id: I74777c793d0c4c7c65f2de927d76bbaac789c001
Signed-off-by: Andrew Walbran <qwandor@google.com>