feat(rust): inverted control flow and service instances
This patch introduces an inverted control flow logic. Instead of calling
exception handlers from lower ELs and "breaking" the EL3 code flow at
each exception return, entering a lower EL is now treated as a function
call which returns when returning to EL3 for any reason (e.g. SMC call,
interrupt, etc.). This allows having a continuous loop that handles all
incoming requests from lower ELs, processes the information, switches to
another world if necessary and enters a lower EL with the appropriate
response. The various services now have instances that encapsulate the
state of the service, and methods that can access it.

Change-Id: Icf1071bd41d0592f7fe7c2f855ed622d7f80212b
Co-authored-by: Andrew Walbran <qwandor@google.com>
Co-authored-by: Imre Kis <imre.kis@arm.com>
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>