feat(rust): add initial interrupt handling
This patch demonstrates how the inverted control flow logic can be used
with asynchronous interrupts. It only covers one scenario: when running
in non-secure state, a secure interrupt will be routed to EL3 and gets
forwarded by the SPMD to the SPMC. Further routing models and interrupt
types can be added later.

Change-Id: I780edcd7b4f708101e4e2607747fd5a1561bf3aa
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>