chore(rust): adopt Linux Foundation's guidance on copyrights
Right now, we advocate for each source file with copyrightable content
to embed a copyright notice of the following format:

  Copyright (c) [XXXX–]YYYY, <Owner>. All rights reserved.

where `[XXXX–]YYYY` are the dates the files was updated. Keeping these
dates accurate over time is a maintenance burden. Also, it unnecessarily
pollutes the git history and can generate meaningless git conflicts.

Adopt the Linux Foundation guidance on this [1], which greatly
simplifies this matter. Not only does it omit the year range but it also
gathers all contributors under a generic "The Rusted Firmware-A
Contributors" term, which avoids the need for each company to add their
own copyright notice. With this new format, the copyright notice
becomes:

  Copyright The Rusted Firmware-A Contributors.

Update the copyright notice requirements in RF-A's style guide.

Also update all rust, assembly and linker script source files in this
regard. Assembly files get treated slightly differently: as these were
inherited from the TF-A project, with possibly contributions from
multiple individuals or companies over the years, the existing
copyright is retained but the new one is added on top of it to cover
recent (and future) contributions.

[1] https://www.linuxfoundation.org/blog/blog/copyright-notices-in-open-source-software-projects

Change-Id: Ib9fb81a53f212aba784134a93b40d07a500b660a
Signed-off-by: Sandrine Afsa <sandrine.afsa@arm.com>