build(rust): use NEED_BL31=no for tf-a's build system
We are currently using tf-a's build system to build bl1, bl2, fip.
We are afterwards updating the fip with the Rust bl31 that we build
with rf-a's build system.

We should therefore avoid build C tf-a's bl31 as we don't need it.

With the current build command presented in the README, we are
currently building c tf-a's bl31 when triggering the fip build.

 * Update README.md to let the user know they should avoid building
   c tf-a's bl31.

Change-Id: I0a4cde9c937005623087a649e4bd001434486104
Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>