docs(rust): apply Arm Trademark Guidance
The RF-A project uses Arm trademarks throughout source files,
documentation and tests. As per the Arm trademark guidance [1], the
first use of each of these trademarks must be preceded by the Arm word
trademark.

This rule can be met by introducing all Arm trademarks used within the
project in the top level documentation, which in the case of RF-A is
the README.md file. This is considered to be "first use" within the
project as whole, with no necessity to adopt "First Use" form within
individual files throughout the source tree.

[1] https://www.arm.com/company/policies/trademarks/guidelines-trademarks

Change-Id: I6b963c5826d3940dc4ab82c534988536c816db37
Signed-off-by: Sandrine Afsa <sandrine.afsa@arm.com>