feat(rust): handle FFA_VERSION forwarding in STF
When Normal World calls FFA_VERSION, that doesn't query the version of
the EL3 FF-A implementation, but it gets forwarded by the SPMD to the
SPMC instead. If the SPMC is compatible with the queried version then it
should return that, otherwise it can return no version or the highest
version it does support. Implement this functionality in STF.

Change-Id: I45c1b2e8a9bd12df7434ede7b52db663cd6ecfad
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>