test(rust): register tests automatically with a macro
Rather than having to list each test in a match statement, this adds a
macro which adds them to a distributed slice via linker shenanigans. The
macro also keeps track of test names so we can log names rather than
just numbers, and for normal-world tests with secure-world helpers gives
them a closure to call it.

Change-Id: I648bb21e45260d1944964ad8ca394558b7390a3c
Signed-off-by: Andrew Walbran <qwandor@google.com>