test: test PSCI CPU_ON and CPU_OFF in STF
This introduces a mechanism for initialising secondary cores in the BL33
component of STF, and uses it to test the PSCI CPU_ON and CPU_OFF calls
by turning a secondary core on and off. This also required an additional
platform API to map the core index to MPIDR value.

Change-Id: I35fa7d6b4561e8d588975560e1737f463ce639ea
Signed-off-by: Andrew Walbran <qwandor@google.com>