fix(rust): fix cargo test on AArch64 host
Currently the modules containing assembly code are only guarded with
cfg(target_arch = "aarch64"). When running cargo test, the code is built
for the host architecture, which means that on an AArch64 machine these
modules are compiled as well. This shouldn't be the case and causes
errors. Add an extra guard to not build these modules for tests.

Change-Id: Id2f99387bf94f8a43b93e9577d2fdbaa8517b3e9
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>