fix(context): use ubfx in get_security_state
The get_security_state macro assumes the register given as output is
clear. However, using the ubfm instruction inserts new bits into the
pre-existing value held in the supplied output register leading to a
wrong world id returned. Restore the macro to using ubfx in place of
ubfm, similary to how it's currently done in the mainline C TF-A.

Also fix the get_per_world_context invocation to use a register (x8)
different from the ones declared as clobbered for this macro (x9, x10).

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I9a1d7300b29c7a78612f00bf4b55d650173c8eeb