docs: list RF-A's hardware & software requirements
By design, RF-A makes a number of assumptions on the hardware
capabilities and on the software stack it interfaces with. These are
mentioned in various comments throughout the code base but we lack a
central document outlining them.

Add a new document to address this, and link it from the README file.

Issue: #60
Change-Id: Icab9e9b23329a2a0c67335e881f9758f3ec07a70
Signed-off-by: Sandrine Afsa <sandrine.afsa@arm.com>