test: distinguish between PSCI MPIDR values and real MPIDR_EL1 values
They can be different due to the MT and U bits, and this distinction is
important on FVP as the MT bit affects whether the affinity fields are
shifted.

As part of this I've copied a subset of the sysregs module from RF-A.
This should probably go into a separate crate at some point.

Change-Id: Iaaff0101e567e3a7f6a4c7de1b17452519de3bc0
Signed-off-by: Andrew Walbran <qwandor@google.com>