test: use RAII to reset sysregs during unit tests
This change adds an RAII class that resets the SYSREGS when the struct
goes out of scope. This is useful because when unit tests fail they
often abort before resetting the global state of SYSREGS. This can lead
to false failures in subsequent tests. By using the RAII resetter, tests
that fail will correctly reset the global state before proceeding to
subsequent tests.

Change-Id: I8d01ea9af57f0736f2fe0b621099298d23600915
Signed-off-by: Jim Ray <jimray@google.com>