refactor: use new autogenerated arm-sysregs crate
The arm-sysregs crate is now published to crates.io, so we should use
that. This requires some changes due to different names of types, use
of bitflags types rather than primitives for more registers, and the
fact that many write functions are now unsafe. (These should have been
unsafe before, but were fixed as part of the other changes.)

Change-Id: I6573f08603adb25863ad0593239a6e139d4afaf8