feat: more traits and methods for MemoryLogger
Making it `repr(C)` means that it has a defined layout, which helps
with extracting the contents from a RAM dump. Implementing `FromBytes`
means that it is safe to initialise with arbitrary bytes, such as the
value remaining from a previous boot after a warm reboot. In this case
it is important to be able to reset the logger to a clean state.

Change-Id: Icd39f0897dc3c33a16c20dc86c837a21b6fc02a4